script/backend/app/requirements.txt
Flask
boto3
requests

beautifulsoup4
python-dotenv

psycopg2-binary 
flask-sqlalchemy